@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Plan2net\Webp\EventListener;
 
-use Exception;
 use Plan2net\Webp\Converter\Exception\ConvertedFileLargerThanOriginalException;
 use Plan2net\Webp\Converter\Exception\WillNotRetryWithConfigurationException;
 use Plan2net\Webp\Service\Configuration;
@@ -16,15 +15,8 @@ use TYPO3\CMS\Core\Resource\FileInterface;
 use TYPO3\CMS\Core\Resource\ProcessedFile;
 use TYPO3\CMS\Core\Resource\ProcessedFileRepository;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use function sprintf;
-use function strpos;
 
-/**
- * Class AfterFileProcessing
- *
- * @author Wolfgang Klinger <wk@plan2.net>
- */
-class AfterFileProcessing
+final class AfterFileProcessing
 {
     public function __invoke(AfterFileProcessingEvent $event): void
     {
@@ -37,11 +29,11 @@ class AfterFileProcessing
     }
 
     /**
-     * Process a file using the configured adapter to create a webp copy
+     * Process a file using the configured adapter to create a webp copy.
      *
      * @param FileInterface|File $file
      */
-    protected function processFile(
+    private function processFile(
         ProcessedFile $processedFile,
         FileInterface $file,
         string $taskType,
@@ -65,7 +57,7 @@ class AfterFileProcessing
                 $file,
                 $taskType,
                 $configuration + [
-                    'webp' => true
+                    'webp' => true,
                 ]
             );
             // Check if reprocessing is required
@@ -86,9 +78,9 @@ class AfterFileProcessing
             } catch (ConvertedFileLargerThanOriginalException $e) {
                 $logger->warning($e->getMessage());
                 $this->removeProcessedFile($processedFileWebp);
-            } catch (Exception $e) {
+            } catch (\Exception $e) {
                 $logger->error(
-                    sprintf(
+                    \sprintf(
                         'Failed to convert image "%s" to webp with: %s',
                         $processedFile->getIdentifier(),
                         $e->getMessage()
@@ -99,7 +91,7 @@ class AfterFileProcessing
         }
     }
 
-    protected function shouldProcess(string $taskType, ProcessedFile $processedFile): bool
+    private function shouldProcess(string $taskType, ProcessedFile $processedFile): bool
     {
         if ('Image.CropScaleMask' !== $taskType) {
             return false;
@@ -123,14 +115,14 @@ class AfterFileProcessing
         return true;
     }
 
-    protected function needsReprocessing(ProcessedFile $processedFile): bool
+    private function needsReprocessing(ProcessedFile $processedFile): bool
     {
-        return $processedFile->isNew() ||
-            (!$processedFile->usesOriginalFile() && !$processedFile->exists()) ||
-            $processedFile->isOutdated();
+        return $processedFile->isNew()
+            || (!$processedFile->usesOriginalFile() && !$processedFile->exists())
+            || $processedFile->isOutdated();
     }
 
-    protected function isFileInProcessingFolder(ProcessedFile $file): bool
+    private function isFileInProcessingFolder(ProcessedFile $file): bool
     {
         $storage = $file->getStorage();
         if (null === $storage) {
@@ -142,10 +134,10 @@ class AfterFileProcessing
             return false;
         }
 
-        return 0 === strpos($file->getIdentifier(), $processingFolder->getIdentifier());
+        return str_starts_with($file->getIdentifier(), $processingFolder->getIdentifier());
     }
 
-    protected function isStorageLocalAndWritable(ProcessedFile $file): bool
+    private function isStorageLocalAndWritable(ProcessedFile $file): bool
     {
         $storage = $file->getStorage();
 
@@ -157,13 +149,13 @@ class AfterFileProcessing
         return 'Local' === $storage->getDriverType() && $storage->isWritable();
     }
 
-    protected function removeProcessedFile(ProcessedFile $processedFile): void
+    private function removeProcessedFile(ProcessedFile $processedFile): void
     {
         try {
             $processedFile->delete(true);
-        } catch (Exception $e) {
+        } catch (\Exception $e) {
             $logger = GeneralUtility::makeInstance(LogManager::class)->getLogger(__CLASS__);
-            $logger->error(sprintf('Failed to remove processed file "%s": %s',
+            $logger->error(\sprintf('Failed to remove processed file "%s": %s',
                 $processedFile->getIdentifier(),
                 $e->getMessage()
             ));
