@@ -4,13 +4,6 @@ declare(strict_types=1);
 
 namespace Plan2net\Webp\Converter\Exception;
 
-use Exception;
-
-/**
- * Class ConvertedFileLargerThanOriginalException
- *
- * @author Wolfgang Klinger <wk@plan2.net>
- */
-class ConvertedFileLargerThanOriginalException extends Exception
+class ConvertedFileLargerThanOriginalException extends \Exception
 {
 }
