@@ -4,13 +4,6 @@ declare(strict_types=1);
 
 namespace Plan2net\Webp\Converter\Exception;
 
-use Exception;
-
-/**
- * Class WillNotRetryWithConfigurationException
- *
- * @author Wolfgang Klinger <wk@plan2.net>
- */
-class WillNotRetryWithConfigurationException extends Exception
+class WillNotRetryWithConfigurationException extends \Exception
 {
 }
