@@ -5,18 +5,11 @@ declare(strict_types=1);
 namespace Plan2net\Webp\Core\Filter;
 
 use TYPO3\CMS\Core\Resource\Driver\DriverInterface;
-use function strlen;
-use function strpos;
 
-/**
- * Class FileNameFilter
- *
- * @author Wolfgang Klinger <wk@plan2.net>
- */
-class FileNameFilter
+final class FileNameFilter
 {
     /**
-     * Remove webp files from file lists
+     * Remove webp files from file lists.
      */
     public static function filterWebpFiles(
         string $itemName,
@@ -25,7 +18,7 @@ class FileNameFilter
         array $additionalInformation,
         DriverInterface $driverInstance
     ): int {
-        if (strpos($itemIdentifier, '.webp') === strlen($itemIdentifier) - 5) {
+        if (\strpos($itemIdentifier, '.webp') === \strlen($itemIdentifier) - 5) {
             return -1;
         }
 
