@@ -8,10 +8,10 @@ $EM_CONF['webp'] = [
     'author_email' => 'wk@plan2.net',
     'state' => 'stable',
     'author_company' => 'plan2net GmbH',
-    'version' => '4.2.0',
+    'version' => '5.0.0',
     'constraints' => [
         'depends' => [
-            'typo3' => '10.4.0-11.5.99'
+            'typo3' => '12.4.0-12.9.99',
         ],
         'conflicts' => [
         ],
